Add edit bookmark command in bm.sh

+++ .scripts/bm.sh
@@ -27,6 +27,8 @@ get_category(){
 
 pick(){
   category_file=$(get_category)
+  [ -z "$category_file" ] && exit 0
+
   runner=$(head -n 1 "$category_file" | awk -F"$SEP" '{print $2}' 2>/dev/null)
   [ -z "$runner" ] && exit 0
 
@@ -57,25 +59,29 @@ add(){
 
     runner=$(rofi -dmenu -p "Runner:")
     [ -z "$runner" ] && exit 0
-    echo "RUNNER$SEP$runner" >> $category_file
+    echo "RUNNER$SEP$runner" >> $category_file && notify-send "Created category '$category' and added '$name'"
   fi
 
   touch "$category_file"
 
   [ -n "$(cat $category_file | grep "$name$SEP" 2>/dev/null)" ] && notify-send "$name already exists" && exit 0
 
-  echo "$name$SEP$val" >> $category_file
+  echo "$name$SEP$val" >> $category_file && notify-send "Added '$name' to '$category' category"
 }
 
 edit(){
   category_file=$(get_category)
+  [ -z "$category_file" ] && exit 0
 
   choice=$(echo -e "category\nbookmark" | rofi -dmenu -p "What are you editing:")
+  [ -z "$choice" ] && exit 0
 
   case "$choice" in
     category)
       category=$(basename $category_file)
+
       choice=$(echo -e "name\nrunner" | rofi -dmenu -p "Edit:")
+      [ -z "$choice" ] && exit 0
 
       case "$choice" in
         name)
@@ -84,12 +90,11 @@ edit(){
           new_category_file="$BM_DIR/$new_category_name"
 
           if [ -f $new_category_file ]; then
-            notify-send "$new_category_file already exists"
-            exit 0
+            notify-send "'$new_category_file' already exists"
+            exit 1
           fi
 
-          mv $category_file $new_category_file
-          notify-send "Renamed category $category to $new_category_name"
+          mv $category_file $new_category_file && notify-send "Renamed category '$category' to '$new_category_name'"
         ;;
         runner)
           runner=$(head -n 1 "$category_file" | awk -F"$SEP" '{print $2}' 2>/dev/null)
@@ -100,23 +105,56 @@ edit(){
           { echo "RUNNER$SEP$choice"; tail -n +2 "$category_file"; } > "$BM_DIR/temp_file" && mv "$BM_DIR/temp_file" "$category_file"
           notify-send "'$category' RUNNER set to '$choice'"
         ;;
+        *)
+          notify-send "Unknown command $choice"
+          exit 1
+        ;;
       esac
     ;;
     bookmark)
-      notify-send "TODO: bm.sh edit bookmark"
-      exit 0
+      bookmark=$(tail -n +2 "$category_file" | awk -F"$SEP" '{print $1}' | rofi -dmenu -p "Bookmark:" 2>/dev/null)
+      [ -z "$bookmark" ] && exit 0
+
+      choice=$(echo -e "name\nvalue" | rofi -dmenu -p "Edit:")
+      case "$choice" in
+        name)
+          new_bookmark=$(rofi -dmenu -p "Edit name:" -filter "$bookmark")
+          [ -z "$new_bookmark" ] && exit 0
+
+          [ -n "$(tail -n +2 "$category_file" | grep "$new_bookmark$SEP")" ] && notify-send "'$new_bookmark' bookmark already exists" && exit 0
+
+          sed -i "s/$bookmark$SEP/$new_bookmark$SEP/" $category_file && notify-send "'$bookmark' renamed to '$new_bookmark'"
+        ;;
+        value)
+          val=$(tail -n +2 "$category_file"| grep "$bookmark$SEP" | awk -F"$SEP" '{print $2}')
+
+          new_val=$(rofi -dmenu -p "Edit name:" -filter "$val")
+          [ -z "$new_val" ] && exit 0
+
+          sed -i "s|$bookmark$SEP$val|$bookmark$SEP$new_val|" $category_file && notify-send "'$bookmark' set to '$new_val'"
+        ;;
+        *)
+          notify-send "Unknown command $choice"
+          exit 1
+        ;;
+      esac
+    ;;
+    *)
+      notify-send "Unknown command $choice"
+      exit 1
     ;;
   esac
 }
 
 delete(){
   category_file=$(get_category)
+  [ -z "$category_file" ] && exit 0
 
-  choice=$(echo -e "category\nbookmark" | rofi -dmenu -p "What are you deleting")
+  choice=$(echo -e "category\nbookmark" | rofi -dmenu -p "What are you deleting:")
 
   case "$choice" in
     category)
-      choice=$(echo -e "yes\nno" | rofi -dmenu -p "Are you sure you want to delete category '$(basename $category_file)'")
+      choice=$(echo -e "yes\nno" | rofi -dmenu -p "Are you sure you want to delete category '$(basename $category_file)':")
       [ -z "$choice" ] && exit 0
 
       [ $choice == "yes" ] && rm -rf $category_file && notify-send "Deleted category $choice"
@@ -125,13 +163,17 @@ delete(){
       bookmark=$(tail -n +2 "$category_file" | awk -F"$SEP" '{print $1}' | rofi -dmenu -p "Bookmark:" 2>/dev/null)
       [ -z "$bookmark" ] && exit 0
 
-      line_no=$(tail -n +2 "$category_file"| grep -n "$bookmark$SEP" | grep -Po "^\d+" )
+      line_no=$(cat "$category_file"| grep -n "$bookmark$SEP" | grep -Po "^\d+" )
+      [ -z "$line_no" ] && notify-send "Could not determine bookmark line no" && exit 0
 
-
-      choice=$(echo -e "yes\nno" | rofi -dmenu -p "Are you sure you want to delete bookmark '$bookmark'")
+      choice=$(echo -e "yes\nno" | rofi -dmenu -p "Are you sure you want to delete bookmark '$bookmark':")
       [ -z "$choice" ] && exit 0
 
-      [ $choice == "yes" ] && sed -i ""$line_no"d" $category_file && notify-send "Deleted bookmark $choice"
+      [ $choice == "yes" ] && sed -i ""$line_no"d" $category_file && notify-send "Deleted bookmark '$bookmark'"
+    ;;
+    *)
+      notify-send "Unknown command $choice"
+      exit 1
     ;;
   esac
 }
@@ -156,7 +198,10 @@ else
     add) add ;;
     edit) edit ;;
     delete) delete ;;
-    *) exit 1 ;;
+    *)
+      notify-send "Unknown command $choice"
+      exit 1
+    ;;
   esac
 fi
 exit 0
